w0vncserver: Move global parameters to common file
Let's keep all global parameters in one place. This makes it easy to
add parameters that can be used by multiple backends in the future.
